backend/requirements-production.txt
# Production requirements for AdCopySurge Backend - Updated 2025-09-11 for Python 3.12 compatibility
# Core currently installed and tested packages

# FastAPI and ASGI server
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
starlette==0.47.3

# Database
sqlalchemy==2.0.43
psycopg2-binary==2.9.9  # PostgreSQL sync driver
asyncpg==0.29.0  # PostgreSQL async driver for SQLAlchemy async
greenlet==3.2.4

# Authentication and Security - VULNERABILITIES FIXED
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1
PyJWT==2.10.1
# SECURITY: Removed ecdsa (CVE-2024-23342) - using cryptography instead
cryptography==46.0.2  # SECURITY: Secure replacement for ecdsa
rsa==4.9.1
pyasn1==0.6.1
# SECURITY: Additional security packages
bleach==6.1.0  # For input sanitization
slowapi==0.1.9  # For rate limiting

# AI and ML - Updated for Python 3.12 compatibility
openai==1.57.2  # Updated for better compatibility
transformers==4.56.1  # Latest version with Python 3.12 wheels
tokenizers==0.22.0
huggingface-hub>=0.34.4,<1.0.0  # Compatible with transformers 4.56.1
fsspec==2024.10.0  # Updated version
textstat==0.7.10
nltk==3.9.1
safetensors==0.6.2
numpy>=1.26.2,<2.0.0  # Fixed: First version with Python 3.12 wheels, avoid v2 for stability
regex==2025.2.10  # SECURITY: Fixed ReDoS vulnerability
packaging==24.1
tqdm==4.67.1
joblib==1.5.2
filelock==3.19.1
# Note: PyTorch and scikit-learn removed for Railway deployment size constraints
# Emotion analysis will use rule-based fallback mode
# To enable AI emotion analysis locally, install: pip install torch scikit-learn

# HTTP requests
httpx>=0.26,<0.29  # Compatible with both openai 1.57.2 and supabase 2.18.1
httpcore==1.0.9
anyio==4.4.0  # SECURITY: Fixed thread race condition vulnerability
sniffio==1.3.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
requests==2.32.5
urllib3==2.5.0
certifi==2024.7.4
charset-normalizer==3.4.3
idna==3.10

# File processing - Updated for Python 3.12 compatibility - SECURITY FIXED
PyPDF4==3.1.0  # SECURITY: Replaced PyPDF2 to fix CVE-2023-36464
python-docx==1.2.0
docx2txt==0.9
lxml>=5.0.0  # Updated for Python 3.12 wheels
python-magic==0.4.27
pyphen==0.17.2

# Data validation
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
jiter==0.10.0
annotated-types==0.7.0
typing_extensions==4.15.0

# Supabase integration
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
postgrest==1.1.1
realtime==2.7.0
storage3==0.12.1
StrEnum==0.4.15
deprecation==2.1.0
dnspython==2.7.0
websockets==15.0.1

# Email validation
email-validator==2.3.0

# Date handling
python-dateutil==2.9.0.post0

# Configuration and utils
PyYAML==6.0.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
six==1.17.0
typing-inspection==0.4.1

# Blog system dependencies
python-frontmatter==1.1.0
python-slugify==8.0.1
markdown==3.7
markdown-it-py==3.0.0
pymdown-extensions==10.7

# Production essentials - Updated for Python 3.12 compatibility
gunicorn==23.0.0  # Updated production WSGI server
alembic==1.13.0  # For database migrations
psutil>=6.0.0  # Updated for system monitoring
structlog==23.2.0  # For logging
redis==5.0.1  # For caching and background tasks
tenacity==9.0.0  # For retry logic and resilience

# Optional production packages
# celery==5.3.4  # For background tasks
# prometheus-client==0.19.0  # For monitoring
# sentry-sdk[fastapi]==1.40.0  # For error tracking
reportlab==4.0.7  # For PDF generation
# pandas==2.1.4  # For data analysis

# Testing (for CI/CD)
# pytest==7.4.3
# pytest-asyncio==0.21.1
